Remove zpagesPipeline interface
